Add fallback for custom archive formats

+++ Pareto Updater/Extensions/PkgApps.swift
@@ -29,13 +29,13 @@ class PkgApp: AppUpdater {
         try? Path(tempPath.path)?.delete()
 
         do {
-            Process.run(app: "/usr/sbin/pkgutil", args: [
+            let pkgRes = Process.run(app: "/usr/sbin/pkgutil", args: [
                 "--expand",
                 appFile.path,
                 tempPath.path
 
             ])
-            Process.run(app: "/usr/bin/tar", args: [
+            let tarRes = Process.run(app: "/usr/bin/tar", args: [
                 "xf",
                 tempPath
                     .appendingPathComponent(pkg, isDirectory: true)
@@ -45,43 +45,54 @@ class PkgApp: AppUpdater {
                 "-C",
                 tempPath.path
             ])
-
-            let app = tempPath.appendingPathComponent(pkgAppName)
-            if let downloadedAppBundle = Bundle(url: app) {
-                if let installedAppBundle = Bundle(url: applicationPath) {
-                    if !validate(downloadedAppBundle, installedAppBundle) {
-                        os_log("Failed to validate app bundle %{public}s", appBundle)
-                        return AppUpdaterStatus.Failed
-                    }
-
-                    do {
-                        os_log("Delete installedAppBundle: \(installedAppBundle.path.string)")
-                        try installedAppBundle.path.delete()
-                        os_log("Update installedAppBundle: \(installedAppBundle.description) with \(downloadedAppBundle.description)")
-                        try downloadedAppBundle.path.copy(to: installedAppBundle.path, overwrite: true)
-                    } catch {
-                        os_log("Delete installedAppBundle failed: \(installedAppBundle.path.string) \(error.localizedDescription)")
-                        if !Process.runCMDasAdmin(cmd: "/bin/rm -rf \(installedAppBundle.path.string)") {
-                            os_log("Delete installedAppBundle failed using admin: \(installedAppBundle.path.string)")
-                            return AppUpdaterStatus.Failed
-                        }
-                        os_log("Update installedAppBundle: \(installedAppBundle.description) with \(downloadedAppBundle.description)")
-                        try downloadedAppBundle.path.copy(to: installedAppBundle.path, overwrite: true)
-                    }
-
-                    if needsStart {
-                        installedAppBundle.launch()
-                    }
-                } else {
-                    os_log("Install AppBundle \(downloadedAppBundle.description)")
-                    try downloadedAppBundle.path.copy(to: Path(applicationPath.path)!, overwrite: true)
+            if tarRes.contains("Unrecognized archive format") {
+                if !Process.runCMDasAdmin(cmd: "/usr/sbin/installer -pkg \(appFile.path) -target /") {
+                    os_log("Install of pkg failed using admin: \(appFile.path)")
+                    return AppUpdaterStatus.Failed
                 }
-
                 if let bundle = Bundle(url: applicationPath), needsStart {
                     bundle.launch()
                 }
                 try? Path(tempPath.path)?.delete()
                 return AppUpdaterStatus.Installed
+            } else {
+                let app = tempPath.appendingPathComponent(pkgAppName)
+                if let downloadedAppBundle = Bundle(url: app) {
+                    if let installedAppBundle = Bundle(url: applicationPath) {
+                        if !validate(downloadedAppBundle, installedAppBundle) {
+                            os_log("Failed to validate app bundle %{public}s", appBundle)
+                            return AppUpdaterStatus.Failed
+                        }
+
+                        do {
+                            os_log("Delete installedAppBundle: \(installedAppBundle.path.string)")
+                            try installedAppBundle.path.delete()
+                            os_log("Update installedAppBundle: \(installedAppBundle.description) with \(downloadedAppBundle.description)")
+                            try downloadedAppBundle.path.copy(to: installedAppBundle.path, overwrite: true)
+                        } catch {
+                            os_log("Delete installedAppBundle failed: \(installedAppBundle.path.string) \(error.localizedDescription)")
+                            if !Process.runCMDasAdmin(cmd: "/bin/rm -rf \(installedAppBundle.path.string)") {
+                                os_log("Delete installedAppBundle failed using admin: \(installedAppBundle.path.string)")
+                                return AppUpdaterStatus.Failed
+                            }
+                            os_log("Update installedAppBundle: \(installedAppBundle.description) with \(downloadedAppBundle.description)")
+                            try downloadedAppBundle.path.copy(to: installedAppBundle.path, overwrite: true)
+                        }
+
+                        if needsStart {
+                            installedAppBundle.launch()
+                        }
+                    } else {
+                        os_log("Install AppBundle \(downloadedAppBundle.description)")
+                        try downloadedAppBundle.path.copy(to: Path(applicationPath.path)!, overwrite: true)
+                    }
+
+                    if let bundle = Bundle(url: applicationPath), needsStart {
+                        bundle.launch()
+                    }
+                    try? Path(tempPath.path)?.delete()
+                    return AppUpdaterStatus.Installed
+                }
             }
             return AppUpdaterStatus.Failed
         } catch {
